+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
-import App from './global/App';
+import App from './ui/App';
 import * as serviceWorker from './config/serviceWorker';
 
 ReactDOM.render(
