+++ src/ui/templates/components/Header/TextLogo.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Text } from '../Base/Base';
+import { Text } from '../../../components/Base/Base';
 
 export default function TextLogo() {
     return (
